Add files via upload
These litter breakdown files that were updated for the manuscript revision. The only change that has occurred between these files and the files in the original folder is that we removed one value from the YR1 coarse-mesh breakdown data. This value was removed due to its implausibility as a very large, negative value for percent AFDM remaining