Add files via upload
This folder contains many redundant files with the original "01_Litter_Breakdown_Data_Carpentry" folder, but some Rmds and .csvs have been updated to reflect changes made during manuscript revision. Specifically, we identified and excluded one very large value for percent AFDM remaining from the YR1 coarse-mesh data, which we are confident was a result of a data processing or data entry error (%AFDM remaining was >300% of initial AFDM). The files in this folder marked with "3Oct23" or "UPD3Oct23" are the ones that have been updated. Analyses in the "05b" folder reflect the removal of this value, which did not change any of our Ea values or our results.