Add files via upload
This is the most up-to-date version of the serial incubations litter breakdown analyses, reflecting changes in which we fixed an error that led to erroneous temperature data being included on 5-6 April 2018. See "01_README_Changes_Since_8May_version.docx" for more details. These changes constitute extremely minor updates to some values (i.e., changes of approximately plus-or-minus 0.01), but do not change the results of the analysis or their implications. However, this is the most up-to-date version of the analysis as of 16 March 2024 and should be used going forward.